backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pandas==2.1.3
openpyxl==3.1.2
aiohttp==3.9.1
pydantic==2.5.0
python-dotenv==1.0.0